+++ api/base/index.ts
@@ -1,10 +1,10 @@
 // Copyright 2017-2023 @polkadot/api authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import type { SignerPayloadRawBase } from 'https://deno.land/x/polkadot@0.2.24/types/types/index.ts';
+import type { SignerPayloadRawBase } from 'https://deno.land/x/polkadot/types/types/index.ts';
 import type { ApiOptions, ApiTypes, DecorateMethod, Signer } from '../types/index.ts';
 
-import { isString, objectSpread, u8aToHex, u8aToU8a } from 'https://deno.land/x/polkadot@0.2.24/util/mod.ts';
+import { isString, objectSpread, u8aToHex, u8aToU8a } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 import { Getters } from './Getters.ts';
 
@@ -26,7 +26,7 @@ export abstract class ApiBase<ApiType extends ApiTypes> extends Getters<ApiType>
    * <BR>
    *
    * ```javascript
-   * import Api from 'https://deno.land/x/polkadot@0.2.24/api/promise/index.ts';
+   * import Api from 'https://deno.land/x/polkadot/api/promise/index.ts';
    *
    * const api = new Api().isReady();
    *