+++ api/types/storage.ts
@@ -1,4 +1,4 @@
 // Copyright 2017-2023 @polkadot/api-augment authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-export * from 'https://deno.land/x/polkadot@0.2.24/api-base/types/storage.ts';
+export * from 'https://deno.land/x/polkadot/api-base/types/storage.ts';