+++ types/checkTypes.manual.ts
@@ -1,13 +1,13 @@
 // Copyright 2017-2023 @polkadot/types authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import 'https://deno.land/x/polkadot@0.2.24/types-augment/mod.ts';
+import 'https://deno.land/x/polkadot/types-augment/mod.ts';
 
-import type { Bytes, Compact, Option, u32 } from 'https://deno.land/x/polkadot@0.2.24/types-codec/mod.ts';
-import type { IOption, ITuple } from 'https://deno.land/x/polkadot@0.2.24/types-codec/types/index.ts';
+import type { Bytes, Compact, Option, u32 } from 'https://deno.land/x/polkadot/types-codec/mod.ts';
+import type { IOption, ITuple } from 'https://deno.land/x/polkadot/types-codec/types/index.ts';
 import type { AccountId, BlockAttestations, SessionKeys7 } from './interfaces/index.ts';
 
-import { assert } from 'https://deno.land/x/polkadot@0.2.24/util/mod.ts';
+import { assert } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 import { TypeRegistry } from './create/index.ts';