+++ types/types/augmentLookup.ts
@@ -3,15 +3,15 @@
 
 // import type lookup before we augment - in some environments
 // this is required to allow for ambient/previous definitions
-import 'https://deno.land/x/polkadot@0.2.24/types-create/types/lookup.ts';
+import 'https://deno.land/x/polkadot/types-create/types/lookup.ts';
 
-import type { Vec } from 'https://deno.land/x/polkadot@0.2.24/types-codec/mod.ts';
-import type { LookupString } from 'https://deno.land/x/polkadot@0.2.24/types-codec/types/index.ts';
-import type { TypeDef } from 'https://deno.land/x/polkadot@0.2.24/types-create/types/index.ts';
+import type { Vec } from 'https://deno.land/x/polkadot/types-codec/mod.ts';
+import type { LookupString } from 'https://deno.land/x/polkadot/types-codec/types/index.ts';
+import type { TypeDef } from 'https://deno.land/x/polkadot/types-create/types/index.ts';
 import type { PortableType } from '../interfaces/metadata/index.ts';
 import type { SiLookupTypeId, SiType } from '../interfaces/scaleInfo/index.ts';
 
-declare module 'https://deno.land/x/polkadot@0.2.24/types-create/types/lookup.ts' {
+declare module 'https://deno.land/x/polkadot/types-create/types/lookup.ts' {
   interface ILookup {
     readonly names: string[];
     readonly types: Vec<PortableType>;